now the value of Digit is stored as a float

+++ src/Calculator/Parser/Digit.java
@@ -2,8 +2,8 @@ package Calculator.Parser;
 import Calculator.Tokenizer.Token;
 
 public class Digit extends TokenValue {
-    public int value;
-    public Digit(Token token, int value){
+    public float value;
+    public Digit(Token token, float value){
         super(token);
         this.value = value;
     }
